+++ backend/views/layouts/main.php
@@ -171,9 +171,13 @@ use yii\widgets\Breadcrumbs;
                         'options'=>['class'=>'treeview'],
                         'items'=>[
                             [
-                                'label'=>Yii::t('backend', 'System Information'),
-                                'url'=>['/system-information/index'],
-                                'icon'=>'<i class="fa fa-angle-double-right"></i>'
+                                'label'=>Yii::t('backend', 'i18n'),
+                                'icon'=>'<i class="fa fa-flag"></i>',
+                                'options'=>['class'=>'treeview'],
+                                'items'=>[
+                                    ['label'=>Yii::t('backend', 'i18n Source Message'), 'url'=>['/i18n/i18n-source-message/index'], 'icon'=>'<i class="fa fa-angle-double-right"></i>'],
+                                    ['label'=>Yii::t('backend', 'i18n Message'), 'url'=>['/i18n/i18n-message/index'], 'icon'=>'<i class="fa fa-angle-double-right"></i>'],
+                                ]
                             ],
                             ['label'=>Yii::t('backend', 'Key-Value Storage'), 'url'=>['/key-storage/index'], 'icon'=>'<i class="fa fa-angle-double-right"></i>'],
                             ['label'=>Yii::t('backend', 'File Storage Items'), 'url'=>['/file-storage/index'], 'icon'=>'<i class="fa fa-angle-double-right"></i>'],
@@ -184,6 +188,11 @@ use yii\widgets\Breadcrumbs;
                                 'icon'=>'<i class="fa fa-angle-double-right"></i>',
                                 'badge'=>\backend\models\SystemEvent::find()->today()->count(),
                                 'badgeBgClass'=>'bg-green',
+                            ],
+                            [
+                                'label'=>Yii::t('backend', 'System Information'),
+                                'url'=>['/system-information/index'],
+                                'icon'=>'<i class="fa fa-angle-double-right"></i>'
                             ],
                             [
                                 'label'=>Yii::t('backend', 'Logs'),