+++ common/config/messages/_base.php
@@ -1,10 +1,9 @@
 <?php
-
-return [
+return  [
     'sourcePath'=>Yii::getAlias('@base'),
     // array, required, list of language codes that the extracted messages
     // should be translated to. For example, ['zh-CN', 'de'].
-    'languages' => ['ru', 'uk'],
+    'languages' => ['uk','ru'],
     // string, the name of the function for translating messages.
     // Defaults to 'Yii::t'. This is used as a mark to find the messages to be
     // translated. You may use a string for single function name or an array for
@@ -16,7 +15,7 @@ return [
     'sort' => false,
     // boolean, whether to remove messages that no longer appear in the source code.
     // Defaults to false, which means each of these messages will be enclosed with a pair of '@@' marks.
-    'removeUnused' => false,
+    'removeUnused' => true,
 
     // array, list of patterns that specify which files/directories should NOT be processed.
     // If empty or not set, all files/directories will be processed.
@@ -44,11 +43,4 @@ return [
         '/vendor',
         '/storage'
     ],
-
-    // 'php' output format is for saving messages to php files.
-    'format' => 'php',
-    // Root directory containing message translations.
-    'messagePath' => Yii::getAlias('@common/messages'),
-    // boolean, whether the message file should be overwritten with the merged messages
-    'overwrite' => true,
 ];