+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Yii Framework 2 Change Log
 1.30 under development
 -----------------------
 
+- Enh: message configs for db, php and po formats
+- Enh: `MessageController` migrate action
+- Enh: I18N CRUD module
+- Enh: `common\components\action\SetLocale`
 - Enh: backendUrlManager, frontendUrlManager, bootstrap application
 
 ... enhancements, bugfixes